# feat: add generic TypedSet with 2x performance boost and compile-time