- Compile-time type safety: No runtime type checking required
- Better performance: Reduced overhead from type assertions and reflection
- Simplified API: Single Add method replaces multiple type-specific methods
- Backward compatibility: Existing Set implementation remains unchanged


benchmark:
`goos: darwin
goarch: arm64
pkg: github.com/zeromicro/go-zero/core/collection
cpu: Apple M2 Pro
BenchmarkTypedIntSet-10         13094462               127.6 ns/op            33 B/op          0 allocs/op
BenchmarkTypedStringSet-10      34456252                33.28 ns/op            6 B/op          1 allocs/op
BenchmarkTypedSetInt-10         16142239               149.8 ns/op            49 B/op          0 allocs/op
BenchmarkRawSet-10               8121650               262.7 ns/op            92 B/op          1 allocs/op
BenchmarkUnmanagedSet-10         7574040               275.9 ns/op            98 B/op          1 allocs/op
BenchmarkSet-10                  7402688               263.4 ns/op           100 B/op          1 allocs/op`